requirements.txt
streamlit
langchain
langchain-community
langchain-google-genai
langchain-huggingface
sentence-transformers
faiss-cpu
beautifulsoup4
python-dotenv
requests
